Move ResourceInformation abstract base class to FWCore/AbstractServices
