+++ FWCore/AbstractServices/BuildFile.xml
@@ -1,0 +1,3 @@
+<export>
+  <lib name="1"/>
+</export>
