+++ FWCore/AbstractServices/src/ResourceInformation.cc
@@ -1,0 +1,6 @@
+#include "FWCore/AbstractServices/interface/ResourceInformation.h"
+
+namespace edm {
+  ResourceInformation::ResourceInformation() = default;
+  ResourceInformation::~ResourceInformation() = default;
+}  // namespace edm
