+++ FWCore/Framework/BuildFile.xml
@@ -4,6 +4,7 @@
 <use name="tbb"/>
 <use name="DataFormats/Common"/>
 <use name="DataFormats/Provenance"/>
+<use name="FWCore/AbstractServices"/>
 <use name="FWCore/Common"/>
 <use name="FWCore/Concurrency"/>
 <use name="FWCore/MessageLogger"/>
