--- FWCore/Utilities/src/ResourceInformation.cc
@@ -1,4 +1,0 @@
-#include "FWCore/Utilities/interface/ResourceInformation.h"
-
-edm::ResourceInformation::ResourceInformation() {}
-edm::ResourceInformation::~ResourceInformation() {}
