+++ HeterogeneousCore/CUDAServices/plugins/BuildFile.xml
@@ -3,6 +3,7 @@
   <use name="tbb"/>
   <use name="DataFormats/Common"/>
   <use name="DataFormats/Provenance"/>
+  <use name="FWCore/AbstractServices"/>
   <use name="FWCore/MessageLogger"/>
   <use name="FWCore/ParameterSet"/>
   <use name="FWCore/ServiceRegistry"/>
