+++ PhysicsTools/PyTorch/test/testBaseCUDA.h
@@ -11,6 +11,7 @@
 #include <cppunit/extensions/HelperMacros.h>
 #include <stdexcept>
 
+#include "FWCore/AbstractServices/interface/ResourceInformation.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
 #include "FWCore/ParameterSetReader/interface/ParameterSetReader.h"
@@ -20,7 +21,6 @@
 #include "FWCore/ServiceRegistry/interface/ServiceRegistry.h"
 #include "FWCore/ServiceRegistry/interface/ServiceToken.h"
 #include "FWCore/Utilities/interface/Exception.h"
-#include "FWCore/Utilities/interface/ResourceInformation.h"
 
 class testBasePyTorchCUDA : public CppUnit::TestFixture {
 public:
