+++ PhysicsTools/TensorFlow/BuildFile.xml
@@ -1,5 +1,6 @@
 <use name="tbb"/>
 <use name="tensorflow-cc"/>
+<use name="FWCore/AbstractServices"/>
 <use name="FWCore/Framework"/>
 <use name="FWCore/MessageLogger"/>
 <use name="FWCore/Utilities"/>
